sndio: Add hack to allow one to overwrite the aucat_cookie path
The path can be set via the AUCAT_COOKIE_PATH env variable.

This can be useful when chaining system user sndiod instances (running
with priority -20) and wanting to access them from different machines
or users at the same time.

This is read-only and if the cookie doesn't exist yet it will not be
generated and a random cookie will be used instead.